backend/requirements.txt
fastapi
uvicorn
python-dotenv
psycopg2-binary
sqlalchemy
pydantic
requests
python-jose
passlib[bcrypt]
python-multipart
bcrypt==4.0.1
pytest
httpx
pytest-asyncio
